requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
jinja2
Babel
python-multipart
itsdangerous
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
mailersend
pillow

# Dependências de teste
pytest
pytest-asyncio
pytest-cov
faker
